requirements.txt
Flask==1.1.1
numpy==1.17.4
python-socketio==4.3.1
requests==2.22.0
tensorflow==2.0.0
# tensorflow-gpu==2.0.0
urllib3==1.25.6
Pillow
opencv-python
